% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiBib.R
\name{EpiBib_AU_YE}
\alias{EpiBib_AU_YE}
\title{TitleEpiBib_AU_YE}
\usage{
EpiBib_AU_YE(author = NULL, year = NULL)
}
\arguments{
\item{author}{The name of the author}

\item{year}{A year}
}
\value{
dataframe with only the references including the selected author and year.
}
\description{
@description This function allows you to search references by author and year.
}
\examples{
df <- EpiBib_AU_YE(author = "AHMAD", year = 2020)
df <- EpiBib_AU_YE(author = "yang")
ddf <- EpiBib_AU_YE(year = 2005)
}
