% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiBib.R
\name{EpiBib_author}
\alias{EpiBib_author}
\title{EpiBib_author}
\usage{
EpiBib_author(author)
}
\arguments{
\item{author}{The name of the author}
}
\value{
dataframe with only the references from the selected author
}
\description{
This function allows you to search references by authors.
}
\examples{
df <- EpiBib_author("Yang")
df <- EpiBib_author("Ya")
df <- EpiBib_author("Yang")
}
