% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiBib.R
\name{EpiBib_country}
\alias{EpiBib_country}
\title{deprecated-EpiBib_country}
\usage{
EpiBib_country(country)
}
\arguments{
\item{country}{The name of the country}
}
\value{
dataframe with only the references from the selected country.
}
\description{
This function allows you to search references by authors' country of origin.
Be careful. countries can be identified differently. For exemple, United Kingdom vs England.
}
\examples{
df <- EpiBib_country("Canada")
df <- EpiBib_country("china")
df <- EpiBib_country("England")
}
