% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiBib_data.R
\docType{data}
\name{EpiBib_data}
\alias{EpiBib_data}
\title{Bibliographic Data on Covid-19}
\format{
A data frame with 26219 rows and 22 variables:
\describe{
 
  ...
}
}
\source{
\url{(https://www.ncbi.nlm.nih.gov/pubmed/}
}
\usage{
EpiBib_data
}
\description{
A dataset containing more than 25 000 references on Covid-19
}
\keyword{datasets}
