% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiBib.R
\name{EpiBib_journal}
\alias{EpiBib_journal}
\title{EpiBib_journal}
\usage{
EpiBib_journal(journal)
}
\arguments{
\item{journal}{The name of the journal}
}
\value{
dataframe with only the references including the selected journal.
}
\description{
This function allows you to search references by journal.
Be careful, journals' name are only avalailable in an abbreviated form.
}
\examples{
df <- EpiBib_journal("med")
df <- EpiBib_journal("bio")
}
