% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiBib.R
\name{EpiBib_references}
\alias{EpiBib_references}
\title{EpiBib_references}
\usage{
EpiBib_references()
}
\value{
The complete bibliographic dataframe
}
\description{
EpiBib_references
}
\examples{
df <- EpiBib_references()
}
