% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiBib.R
\name{EpiBib_source}
\alias{EpiBib_source}
\title{deprecated-EpiBib_source}
\usage{
EpiBib_source(source)
}
\arguments{
\item{source}{The name of the source}
}
\value{
dataframe with only the references including the selected source
}
\description{
This function allows you to search references by source
Be careful, souces' name are only avalailable in an abbreviated form.
}
\examples{
df <- EpiBib_source("med")
df <- EpiBib_source("bio")
}
