% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiBib.R
\name{EpiBib_title}
\alias{EpiBib_title}
\title{EpiBib_title}
\usage{
EpiBib_title(title)
}
\arguments{
\item{title}{keyword in the title}
}
\value{
dataframe with only the references including the selected word in the title.
}
\description{
This function allows you to search references by keywords in the title.
}
\examples{
df <- EpiBib_title("pandemic")
df <- EpiBib_title("coronaVir")
}
