% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiBib.R
\name{EpiBib_year}
\alias{EpiBib_year}
\title{EpiBib_year}
\usage{
EpiBib_year(year)
}
\arguments{
\item{year}{A year}
}
\value{
dataframe with only the references including the selected year.
}
\description{
This function allows you to search references by year.
}
\examples{
df <- EpiBib_year("2017")
}
