% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiBib.R
\name{epibibr_visual}
\alias{epibibr_visual}
\title{epibibr_visual}
\usage{
epibibr_visual(chart = "line_1", title = TRUE)
}
\arguments{
\item{chart}{Type of charts}

\item{title}{Chart title, if TRUE will appear}
}
\value{
Graph chosen
}
\description{
This function allows you to create visuals from the preset options :
 line chart on number of articles, bar chart on most productive authors or bar chart on most productive countries.
}
