% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiBib.R
\name{update_dataset}
\alias{update_dataset}
\title{update_dataset}
\usage{
update_dataset()
}
\value{
An updated dataset
}
\description{
this function allows you to update your dataset
}
\examples{
update_dataset()
}
