% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiBib.R
\name{epibib_data}
\alias{epibib_data}
\title{epibib_data}
\usage{
epibib_data(
  author = "",
  year = "",
  country = "",
  title = "",
  source = "",
  abstract = ""
)
}
\arguments{
\item{author}{The name of an author}

\item{year}{Years for which you want the data.}

\item{country}{Author's country of origin}

\item{title}{Keyword in the title}

\item{source}{Name of the source}

\item{abstract}{Keyword in the abstract}
}
\value{
Data for the author, country, year, title, source and abstract requested
}
\description{
this function allows you to to find and display medical bibliographic data according to the selected parameters.
If no arguments are filled, all data will be displayed.
}
\examples{
EpiBib_data <- epibib_data()
}
